+++ lib/hexapdf/type/acro_form/variable_text_field.rb
@@ -154,13 +154,21 @@ module HexaPDF
         # font_size, font_color].
         #
         # The default appearance string is taken from the given +widget+ of the field, falls back to
-        # the field itself or, if still not available, the default appearance string of the form.
+        # the field itself and then the default appearance string of the form. If it still not
+        # available, a standard default appearance string is set (see
+        # #set_default_appearance_string) and used.
         #
         # The reason why a specific widget of the field can be specified is because the widgets of a
         # field might differ in their visual representation.
         def parse_default_appearance_string(widget = self)
           da = widget[:DA] || self[:DA] || (document.acro_form && document.acro_form[:DA])
-          raise HexaPDF::Error, "No default appearance string set" unless da
+          unless da
+            if (args = document.config['acro_form.fallback_default_appearance'])
+              da = set_default_appearance_string(**args)
+            else
+              raise HexaPDF::Error, "No default appearance string set"
+            end
+          end
           self.class.parse_appearance_string(da)
         end